requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pinecone==3.2.2
openai==1.3.7
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
numpy==1.26.4
sentence-transformers==2.2.2
requests==2.31.0
pydantic==2.5.0